backend/requirements.txt
asgiref==3.11.0
django==6.0
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
pillow==12.0.0
pyjwt==2.10.1
sqlparse==0.5.5
